Mirror changes from master.

+++ cups/ppd-cache.c
@@ -3130,22 +3130,22 @@ _ppdCreateFromIPP(char   *buffer,	/* I - Filename buffer */
 					/* Type of password */
 
     if (maxlen > (int)(sizeof(pattern) - 1))
-      maxlen = sizeof(pattern) - 1;
+      maxlen = (int)sizeof(pattern) - 1;
 
     if (!repertoire || !strcmp(repertoire, "iana_us-ascii_digits"))
-      memset(pattern, '1', maxlen);
+      memset(pattern, '1', (size_t)maxlen);
     else if (!strcmp(repertoire, "iana_us-ascii_letters"))
-      memset(pattern, 'A', maxlen);
+      memset(pattern, 'A', (size_t)maxlen);
     else if (!strcmp(repertoire, "iana_us-ascii_complex"))
-      memset(pattern, 'C', maxlen);
+      memset(pattern, 'C', (size_t)maxlen);
     else if (!strcmp(repertoire, "iana_us-ascii_any"))
-      memset(pattern, '.', maxlen);
+      memset(pattern, '.', (size_t)maxlen);
     else if (!strcmp(repertoire, "iana_utf-8_digits"))
-      memset(pattern, 'N', maxlen);
+      memset(pattern, 'N', (size_t)maxlen);
     else if (!strcmp(repertoire, "iana_utf-8_letters"))
-      memset(pattern, 'U', maxlen);
+      memset(pattern, 'U', (size_t)maxlen);
     else
-      memset(pattern, '*', maxlen);
+      memset(pattern, '*', (size_t)maxlen);
 
     pattern[maxlen] = '\0';
 
@@ -3866,7 +3866,7 @@ _ppdCreateFromIPP(char   *buffer,	/* I - Filename buffer */
         if (tray_len >= (int)sizeof(tray))
           tray_len = (int)sizeof(tray) - 1;
 
-        memcpy(tray, tray_ptr, tray_len);
+        memcpy(tray, tray_ptr, (size_t)tray_len);
         tray[tray_len] = '\0';
 
         if (strstr(tray, "stackingorder=lastToFirst;"))
